requirements.txt
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
django-heroku==0.3.1
django-dotenv==1.4.2
django-filter==2.2.0
django-sendgrid-v5==0.8.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
future==0.18.2
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
python-dotenv==0.10.3
python-http-client==3.2.1
pytz==2019.3
sendgrid==6.1.0
sqlparse==0.3.0
waitress==1.3.1
whitenoise==4.1.4
